New clases AbstractDrive and CombinedDrive; __shift__ to concatenate
